+++ src/app/api/instructors/check-nickname/route.ts
